moving flatten stash tabs to item set manager; modified imports
